--- a/pxd_bio_blkmq.c
+++ b/pxd_bio_blkmq.c
@@ -214,10 +214,11 @@ void __pxd_abortfailQ(struct pxd_device *pxd_dev) {
 // no locking needed, @ios is a local list of IO to be reissued.
 void pxd_reissuefailQ(struct pxd_device *pxd_dev, struct list_head *ios,
                       int status) {
-        while (!list_empty(ios)) {
-                struct fp_root_context *fproot = list_first_entry(
-                    &pxd_dev->fp.failQ, struct fp_root_context, wait);
-                struct fuse_req *req = fproot_to_fuse_request(fproot);
+        struct fp_root_context *fproot;
+        struct fp_root_context *tmp;
+
+        list_for_each_entry_safe(fproot, tmp, ios, wait) {
+                struct fuse_req* req = fproot_to_fuse_request(fproot);
                 BUG_ON(fproot->magic != FP_ROOT_MAGIC);
                 list_del(&fproot->wait);
                 clone_cleanup(fproot);